[FLINK-3943] Fixed according to PR comments
