+++ docs/apis/table.md
@@ -550,7 +550,7 @@ Table result = left.intersect(right);
     <tr>
       <td><strong>Minus</strong></td>
       <td>
-        <p>Similar to a SQL EXCEPT clause. Except returns records from the first table that do not exist in the second table. Duplicate records in the first table are returned exactly once, i.e., duplicates are removed. Both tables must have identical schema, i.e., field names and types.</p>
+        <p>Similar to a SQL EXCEPT clause. Minus returns records from the left table that do not exist in the right table. Duplicate records in the left table are returned exactly once, i.e., duplicates are removed. Both tables must have identical schema, i.e., field types.</p>
 {% highlight java %}
 Table left = tableEnv.fromDataSet(ds1, "a, b, c");
 Table right = tableEnv.fromDataSet(ds2, "a, b, c");
@@ -755,7 +755,7 @@ val result = left.intersect(right);
     <tr>
       <td><strong>Minus</strong></td>
       <td>
-        <p>Similar to a SQL EXCEPT clause. Returns elements from the first table that do not exist in the second table. Both tables must have identical schema(field names and types).</p>
+        <p>Similar to a SQL EXCEPT clause. Returns elements from the left table that do not exist in the right table. Both tables must have identical schema(field types).</p>
 {% highlight scala %}
 val left = ds1.toTable(tableEnv, 'a, 'b, 'c);
 val right = ds2.toTable(tableEnv, 'a, 'b, 'c);
